requirements.txt
Flask==2.3.3
python-pam==1.8.4
vici==5.9.0